corrigindo método de inserir foto
